Add audio filter.
- Implemented an order two butterworth filter.

+++ include/nois/NoisTypes.hpp
@@ -1,8 +1,10 @@
 #pragma once
 
+#include <cinttypes>
+
 namespace nois {
 
 using data_t = float;
-using count_t = int;
+using count_t = int64_t;
 
 };
